Moving exceptions out from internal package, as they are public
